SecurityDirectives passing
Two faulty tests and two actual bugs.
